Create marquee.py
This code queries the API for the information that would show up on the station platform. There have been some modifications made. when the time remaining reads either BRD (boarding) or ARR (arriving) the time part turns red to reflect the urgency. 
Red,Green,Blue and Yellow lines are colored according to the name of the line. Termcolor does not support Orange this may be revisited with a different library.